修复平均值计算错误

+++ slate-common/src/main/java/sviolet/slate/common/utilx/txtimer/def/Reporter.java
@@ -156,7 +156,7 @@ class Reporter {
                     if (info.finishTotal > 100000) {
                         changeRate = 0.03f;
                     } else {
-                        changeRate = info.finish / info.finishTotal;
+                        changeRate = (float)info.finish / (float)info.finishTotal;
                         if (changeRate < 0.03f) {
                             changeRate = 0.03f;
                         }
